Rename FooResponse namespace (#1279)
