+++ src/ErrorResponse/EntityErrorInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\FooResponse;
+namespace App\ErrorResponse;
 
 use App\Entity\IdentifyingEntityInterface;
