+++ src/ErrorResponse/Serializer/ComponentFactory/RequirementsFactory.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace App\FooResponse\ErrorSerializer\ComponentFactory;
+namespace App\ErrorResponse\Serializer\ComponentFactory;
 
-use App\FooResponse\BadRequestErrorInterface;
-use App\FooResponse\ErrorInterface;
-use App\FooResponse\ErrorSerializer\Component;
-use App\FooResponse\ErrorSerializer\ComponentFactoryInterface;
-use App\FooResponse\SizeInterface;
+use App\ErrorResponse\BadRequestErrorInterface;
+use App\ErrorResponse\ErrorInterface;
+use App\ErrorResponse\Serializer\Component;
+use App\ErrorResponse\Serializer\ComponentFactoryInterface;
+use App\ErrorResponse\SizeInterface;
 use App\RequestField\RequirementsInterface;
 use App\RequestField\ScalarRequirementsInterface;
