+++ src/RequestField/Field/ScalarRequirements.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\RequestField\Field;
 
-use App\FooResponse\SizeInterface;
+use App\ErrorResponse\SizeInterface;
 use App\RequestField\ScalarRequirementsInterface;
 
 readonly class ScalarRequirements extends Requirements implements ScalarRequirementsInterface
